requirements.txt
Django==3.1.6
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
djangorestframework_simplejwt
djoser==2.0.5
psycopg2==2.8.6
pytz==2021.1
